refactor advisory policy action routes
