+++ src/app/router_registry.py
@@ -11,6 +11,7 @@ from app.routers.advisor_cockpit_projections import (
     router as advisor_cockpit_projections_router,
 )
 from app.routers.advisory_policy import router as advisory_policy_router
+from app.routers.advisory_policy_actions import router as advisory_policy_actions_router
 from app.routers.advisory_policy_evaluation_actions import (
     router as advisory_policy_evaluation_actions_router,
 )
@@ -249,6 +250,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(advisory_workspace_assistant_router)
     app.include_router(advisory_workspace_handoff_router)
     app.include_router(advisory_policy_router)
+    app.include_router(advisory_policy_actions_router)
     app.include_router(advisory_policy_evaluations_router)
     app.include_router(advisory_policy_evaluation_evidence_router)
     app.include_router(advisory_policy_evaluation_actions_router)
